Move LLVMOption into a separate file
